Use the $ServiceName parameter when installing.
Otherwise TopShelf drops back to what is set in the HostFactory/start-up
Also, check if app.config exists before trying to rename it

+++ topshelf/Deploy.ps1
@@ -23,7 +23,7 @@ if (! $service)
 {
     Write-Host "The service will be installed"
 
-	& "$fullPath" "install" | Write-Host
+	& "$fullPath" install --servicename:$ServiceName | Write-Host
 }
 else
 {
@@ -33,14 +33,15 @@ else
     & "sc.exe" config $service.Name binPath= $fullPath start= auto | Write-Host
 }
 
+if(Test-Path "app.config")
+{
+	# rename the transformed "app.config" to "MyService.exe.config" as this is not handled automatically
+	# by octopus unless your transforms match the service exe config name ("MyService.<Environment>.exe.config")
+	Write-Host "Copying transformed $OctopusEnvironmentName configuration file"
 
-# rename the transformed "app.config" to "MyService.exe.config" as this is not handled automatically
-# by octopus unless your transforms match the service exe config name ("MyService.<Environment>.exe.config")
-Write-Host "Copying transformed $OctopusEnvironmentName configuration file"
-
-Rename-Item "$ServiceExecutable.config" "$ServiceExecutable.config.original"
-Copy-Item "app.config" -Destination "$ServiceExecutable.config"
-
+	Rename-Item "$ServiceExecutable.config" "$ServiceExecutable.config.original"
+	Copy-Item "app.config" -Destination "$ServiceExecutable.config"
+}
 
 # start !
 Write-Host "Starting the service"
